Start to work on the rewrite.
I decided to simplify this because I noticed that when I had trouble
figuring out how to accomplish a goal with it, or to work around a slow
point, it was because at_once was trying to be clever and doing things
for me that I didn't want it to do (such as writing to disk twice to use
the cache, or renaming chunks from input to output).

The basic idea of this rewrite is:
* No more cache.
* No more reduce.
* Pickle-in/out is default. jsonl-in/out can happen if you ask.
  - Inputs are always full chunks. If you want to process item by item,
    just do that in a for loop in your map function.
  - Output keys are only hashed if you specify a number of out shards.

I think I can support all of the earlier functionality still, while
keeping the code and interface simpler. My primary goal is to keep it
flexible and simple while empowering as much speed as Python can
reasonably provide.

If you want to run a traditional map-reduce, you can still do this in
two jobs. The first job outputs to a given number of jsonl shards, and
the second job receives those shards and can reduce them.